requirements.txt
# Generated via Nexin django powered rest API template
# Remember to add specific versions for packages to ensure compatibility

Django==4.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
dj-static==0.0.6
Pillow==9.2.0
django-storages==1.12.3
psycopg2-binary==2.9.3
dj-database-url==1.0.0
pre-commit==2.20.0
black==22.6.0
requests==2.28.1
python-decouple==3.6
Unipath==1.1
drf-yasg==1.21.3
django-on-heroku==1.1.2
gunicorn==20.1.0
boto3==1.24.45
botocore==1.27.45
s3transfer==0.6.0
